Exceções
Coloquei uma no de client e uma no de agencia
