--- apps/worker/template/worker/buscasFuncionario/dadosFuncionario.html
@@ -1,38 +1,0 @@
-<!DOCTYPE html>
-{% load static %}
-
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Title</title>
-    <link href="{% static 'estilos/dados_print.css' %}" rel="stylesheet" type="text/css"/>
-</head>
-<body>
-
-
-
-{% for cadastro in object_list %}
-
-      <div class="dados_cliente">
-
-        <h1>Detalhes do Cadastro</h1>
-         <hr/>
-        <div class="apenasp_cliente">
-
-          <h2>Dados do funcionário</h2>
-         <p><strong>Nome: </strong>{{ cadastro.first_name }} {{ cadastro.last_name }}</p>
-        <p><strong>Email: </strong> {{ cadastro.email }}</p>
-        <p><strong>Telefone: </strong> {{ cadastro.telefone }}</p>
-        <p><strong>Cargo: </strong> {{ cadastro.cargo_atual }}</p>
-         <p><strong>Salario: </strong> {{ cadastro.salario }}</p>
-         <p><strong>Situação: </strong> {{ cadastro.situacao_atual }}</p>
-         <p><strong>Situação na empre sa: </strong> {{ cadastro.situacao_atual_empresa }}</p>
-        </div>
-
-    <button class="btnvoltar"><a href="{% url 'AdminListagemFuncionario' %}">Voltar</a></button>
-{% endfor %}
-
-
-
-</body>
-</html>
