--- apps/worker/template/worker/buscasFuncionario/procurarFuncionario.html
@@ -1,80 +1,0 @@
-<!DOCTYPE html>
-{% load static %}
-
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Resultado da Pesquisa</title>
-    <link href="{% static 'estilos/dados_list.css' %}" rel="stylesheet" type="text/css"/>
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha3/dist/css/bootstrap.min.css" rel="stylesheet">
-</head>
-<body class="bg-light">
-
-<div class="container mt-5">
-    <h1 class="text-center mb-4">Resultado da Pesquisa</h1>
-
-    <div class="row justify-content-center mb-4">
-        <div class="col-md-8">
-            <form action="{% url 'procurarFuncionario' %}" method="GET" class="input-group">
-                <input type="search" class="form-control" name="q" value="{{ procurar_termo }}" required>
-                <button type="submit" class="btn btn-primary">
-                    <img src="{% static 'imagens/lupe.svg' %}" alt="Ícone de lupa" style="height: 20px; width: 20px;">
-                </button>
-            </form>
-        </div>
-    </div>
-
-    {% if total_resultados > 0 %}
-        <div class="results-count mb-4">
-            <h2>Total de resultados: {{ total_resultados }}</h2>
-        </div>
-    {% endif %}
-
-    <div class="table-responsive">
-        <table class="table table-striped table-hover">
-            <thead class="table-dark">
-                <tr>
-                    <th>#</th>
-                    <th>Nome</th>
-                    <th>Email</th>
-                    <th>Telefone</th>
-                    <th>Cargo</th>
-                    <th>Comissão</th>
-                    <th>Situação</th>
-                    <th>Ações</th>
-                </tr>
-            </thead>
-            <tbody>
-                {% for cadastro in object_list %}
-                    <tr>
-                        <td><a href="{% url 'dadosFuncionario' cadastro.pk %}" class="text-decoration-none">{{ cadastro.pk }}</a></td>
-                        <td><a href="{% url 'dadosFuncionario' cadastro.pk %}" class="text-decoration-none">{{ cadastro.first_name }} {{ cadastro.last_name }}</a></td>
-                        <td>{{ cadastro.email }}</td>
-                        <td>{{ cadastro.telefone }}</td>
-                        <td>{{ cadastro.cargo_atual }}</td>
-                        <td>{{ cadastro.comissao_acumulada }}</td>
-                        <td>{{ cadastro.situacao_atual_empresa }}</td>
-                        <td>
-                            <a href="{% url 'funcionario_atualizar' pk=cadastro.pk %}" class="btn btn-secondary btn-sm {% if cadastro.finished_at %}disabled{% endif %}">
-                                <img style="display: inline-block; width: 15px; height: 15px" src="{% static 'imagens/edit.svg' %}" alt="Editar"> Editar
-                            </a>
-                        </td>
-                    </tr>
-                {% empty %}
-                    <tr>
-                        <td colspan="8" class="text-center text-muted">Sem registros</td>
-                    </tr>
-                {% endfor %}
-            </tbody>
-        </table>
-    </div>
-
-    <div class="text-center mt-4">
-        <a href="{% url 'AdminListagemFuncionario' %}" class="btn btn-outline-primary">Voltar para a Pesquisa</a>
-    </div>
-</div>
-
-<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha3/dist/js/bootstrap.bundle.min.js"></script>
-</body>
-</html>
