--- apps/worker/template/worker/formsFuncionario/cadastroFuncionario_confirm_delete.html
@@ -1,30 +1,0 @@
-<!DOCTYPE html>
-{% load static %}
-
-<html lang="en">
-<head>
-   <meta charset="UTF-8">
-   <title> Excluír Tarefa</title>
-        <link href="{% static 'estilos/confirm_delete.css' %}" rel="stylesheet" type="text/css"/>
-    <link href="https://fonts.googleapis.com/css2?family=Roboto:wght@400;500;700&display=swap" rel="stylesheet">
-</head>
-<body>
-
-<div class="container">
-<h1>Excluir Funcionário?</h1>
-<p>Tem certeza que deseja excluir o funcionário <strong>{{ object.first_name }} {{ object.last_name }}</strong>?</p>
-
-
-<form method="POST">
-   {% csrf_token %}
-   <button type="submit" class="btn btn-danger">Excluir</button>
-
-      <a href="{% url 'AdminListagemFuncionario' %}?page={{ request.GET.page }}" class="btn btn-primary">Cancelar</a>
-</form>
-
-</div>
-
-
-</body>
-</html>
-
