--- apps/worker/template/worker/formsFuncionario/cadastroFuncionario_form.html
@@ -1,49 +1,0 @@
-<!DOCTYPE html>
-{% load crispy_forms_filters %}
-{% load crispy_forms_tags %}
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Registro de Funcionário</title>
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-KK94CHFLLe+nY2dmCWGMq91rCGa5gtU4mk92HdvYe+M/SXH301p5ILy+dN9+nJOZ" crossorigin="anonymous">
-</head>
-<body>
-<div class="container mt-5">
-    <div class="right-section">
-        <h2>Registro</h2><br>
-        <form method="POST">
-            {% csrf_token %}
-            {{ form|crispy }}
-            <button type="submit" class="btn btn-primary">Salvar dados do funcionário</button>
-        </form>
-        <br>
-        <p><a class="btn btn-primary" href="{% url 'AdminListagemFuncionario' %}">Voltar</a></p>
-    </div>
-</div>
-
-
-<script>
-    // Javascript para atualizar dinamicamente as opções de 'cargo_atual'
-document.addEventListener('DOMContentLoaded', function() {
-    var situacaoField = document.getElementById('id_situacao_atual');
-    var cargoField = document.getElementById('id_cargo_atual');
-
-    situacaoField.addEventListener('change', function() {
-        var situacao = situacaoField.value;
-
-        if (situacao === "Adm.") {
-            cargoField.innerHTML = '<option value="Financeiro">Financeiro</option>';
-        } else if (situacao === "Exec") {
-            cargoField.innerHTML = '<option value="Executivo contas">Executivo contas</option>';
-        } else if (situacao === "Func.") {
-            cargoField.innerHTML = '<option value="Despachante">Despachante</option>' +
-                                   '<option value="Despachante externo">Despachante externo</option>' +
-                                   '<option value="Suporte Whatsapp">Suporte Whatsapp</option>';
-        } else {
-            cargoField.innerHTML = '<option value="">---</option>';
-        }
-    });
-});
-</script>
-</body>
-</html>
