--- apps/worker/template/worker/formsFuncionario/cadastroFuncionario_list.html
@@ -1,100 +1,0 @@
-<!DOCTYPE html>
-{% load static %}
-
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Listagem de Funcionários</title>
-    <link href="{% static 'estilos/dados_list.css' %}" rel="stylesheet" type="text/css"/>
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha3/dist/css/bootstrap.min.css" rel="stylesheet">
-</head>
-<body class="bg-light">
-
-<div class="container mt-5">
-    <h1 class="text-center mb-4">Listagem de Funcionários</h1>
-
-    <div class="row justify-content-center mb-4">
-        <div class="col-md-8">
-            <form action="{% url 'procurarFuncionario' %}" method="GET" class="input-group">
-                <input type="search" class="form-control" name="q" placeholder="Procure por funcionários..." value="{{ procurar_termo }}" required>
-                <button type="submit" class="btn btn-primary">
-                    <img src="{% static 'imagens/lupe.svg' %}" alt="Ícone de lupa" style="height: 20px; width: 20px;">
-                </button>
-            </form>
-        </div>
-    </div>
-
-    {% if cadastro_list %}
-        <div class="table-responsive">
-            <table class="table table-striped table-hover">
-                <thead class="table-dark">
-                    <tr>
-                        <th>#</th>
-                        <th>Nome</th>
-                        <th>Email</th>
-                        <th>Telefone</th>
-                        <th>Cargo</th>
-                        <th>Comissão</th>
-                        <th>Situação</th>
-                        <th>Ações</th>
-                    </tr>
-                </thead>
-                <tbody>
-                {% for todo in cadastro_list %}
-                    <tr>
-                        <td><a href="{% url 'dadosFuncionario' todo.pk %}" class="text-decoration-none">{{ todo.pk }}</a></td>
-                        <td><a href="{% url 'dadosFuncionario' todo.pk %}" class="text-decoration-none">{{ todo.first_name }} {{ todo.last_name }}</a></td>
-                        <td>{{ todo.email }}</td>
-                        <td>{{ todo.telefone }}</td>
-                        <td>{{ todo.cargo_atual }}</td>
-                        <td>{{ todo.comissao_acumulada }}</td>
-                        <td>{{ todo.situacao_atual_empresa }}</td>
-                        <td class="botoestb">
-                            <a href="{% url 'funcionario_atualizar' pk=todo.pk %}" class="btn btn-secondary btn-sm {% if todo.finished_at %}disabled{% endif %}">
-                                <img style="display: inline-block; width: 15px; height: 15px" src="{% static 'imagens/edit.svg' %}" alt="Editar"> Editar
-                            </a>
-                        </td>
-                    </tr>
-                {% endfor %}
-                </tbody>
-            </table>
-        </div>
-    {% else %}
-        <p class="text-center text-muted">Não há funcionários cadastrados.</p>
-    {% endif %}
-
-    <nav class="d-flex justify-content-center">
-        <ul class="pagination">
-            {% if page_obj.has_previous %}
-                <li class="page-item">
-                    <a class="page-link" href="?page=1&nome={{ request.GET.nome }}">&laquo; Primeira</a>
-                </li>
-                <li class="page-item">
-                    <a class="page-link" href="?page={{ page_obj.previous_page_number }}&nome={{ request.GET.nome }}">Anterior</a>
-                </li>
-            {% endif %}
-
-            <li class="page-item disabled">
-                <span class="page-link">Página {{ page_obj.number }} de {{ page_obj.paginator.num_pages }}</span>
-            </li>
-
-            {% if page_obj.has_next %}
-                <li class="page-item">
-                    <a class="page-link" href="?page={{ page_obj.next_page_number }}&nome={{ request.GET.nome }}">Próxima</a>
-                </li>
-                <li class="page-item">
-                    <a class="page-link" href="?page={{ page_obj.paginator.num_pages }}&nome={{ request.GET.nome }}">Última &raquo;</a>
-                </li>
-            {% endif %}
-        </ul>
-    </nav>
-
-    <div class="text-center mt-4">
-        <a href="{% url 'homeAdmin' %}" class="btn btn-outline-primary">Voltar</a>
-    </div>
-</div>
-
-<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha3/dist/js/bootstrap.bundle.min.js"></script>
-</body>
-</html>
